Define type conversions explicitly (Reference-LAPACK PR 703)

+++ lapack-netlib/SRC/sgglse.f
@@ -276,7 +276,7 @@
 *
       CALL SGGRQF( P, M, N, B, LDB, WORK, A, LDA, WORK( P+1 ),
      $             WORK( P+MN+1 ), LWORK-P-MN, INFO )
-      LOPT = WORK( P+MN+1 )
+      LOPT = INT( WORK( P+MN+1 ) )
 *
 *     Update c = Z**T *c = ( c1 ) N-P
 *                          ( c2 ) M+P-N
